Explicitly list the subdirectories
Globbing isn't good since if we enter PoolAllocate before DSA, there
will be a CMake error since PoolAllocate requires objects from DSA. I
encountered this problem while trying to build on a platform that
visited the directories in this order. By explicitly listing the
directories we can control the order that they will be visited.